+++ Tests/Set-CIisWebsiteLimit.Tests.ps1
@@ -1,0 +1,196 @@
+using module '..\Carbon.Iis'
+using namespace Microsoft.Web.Administration
+
+Set-StrictMode -Version 'Latest'
+
+BeforeAll {
+    & (Join-Path -Path $PSScriptRoot 'Initialize-CarbonTest.ps1' -Resolve)
+
+    $script:testNum = 0
+
+    $script:defaultDefaults = @{}
+    (Get-CIisWebsite -Defaults).limits.Attributes |
+        Where-Object 'IsInheritedFromDefaultValue' -EQ $false |
+        ForEach-Object { $script:defaultDefaults[$_.Name] = $_.Value }
+
+    # All non-default values.
+    $script:nonDefaultArgs = @{
+        'connectionTimeout' = '00:01:00';
+        'maxBandwidth' = 2147483647;
+        'maxConnections' = 1073741823;
+        'maxUrlSegments' = 16;
+    }
+
+    # Values that once set, can only be changed, never removed.
+    $script:requiredDefaults = @{
+    }
+    # Sometimes the default values in the schema aren't quite the default values.
+    $script:notQuiteDefaultValues = @{
+    }
+
+    $script:excludedAttributes = @()
+
+    function ThenDefaultsSetTo
+    {
+        param(
+            [hashtable] $Values = @{},
+
+            [hashtable] $OrValues = @{}
+        )
+
+        ThenHasValues $Values -OrValues $OrValues -OnDefaults
+        ThenHasValues $Values -OrValues $OrValues
+    }
+
+    function ThenHasDefaultValues
+    {
+        ThenHasValues @{}
+    }
+
+    function ThenHasValues
+    {
+        param(
+            [Parameter(Position=0)]
+            [hashtable] $Values = @{},
+
+            [hashtable] $OrValues = @{},
+
+            [switch] $OnDefaults
+        )
+
+        $targetParent = Get-CIisWebsite -Name $script:siteName -Defaults:$OnDefaults
+        $targetParent | Should -Not -BeNullOrEmpty
+
+        $target = $targetParent.limits
+        $target | Should -Not -BeNullOrEmpty
+
+        $asDefaultsMsg = ''
+        if( $OnDefaults )
+        {
+            $asDefaultsMsg = ' as default'
+        }
+
+        foreach( $attr in $target.Schema.AttributeSchemas )
+        {
+            if( $attr.Name -in $script:excludedAttributes )
+            {
+                continue
+            }
+
+            $expectedValue = $attr.DefaultValue
+            $becauseMsg = 'default'
+            $setMsg = 'set'
+            if( $script:notQuiteDefaultValues.ContainsKey($attr.Name))
+            {
+                $expectedValue = $script:notQuiteDefaultValues[$attr.Name]
+            }
+
+            if( $Values.ContainsKey($attr.Name) )
+            {
+                $expectedValue = $Values[$attr.Name]
+                $becauseMsg = 'custom'
+            }
+            elseif( $OrValues.ContainsKey($attr.Name) )
+            {
+                $expectedValue = $OrValues[$attr.Name]
+                $becauseMsg = 'custom'
+                $setMsg = 'not set'
+            }
+
+            $currentValue = $target.GetAttributeValue($attr.Name)
+            if( $currentValue -is [securestring] )
+            {
+                $currentValue = [pscredential]::New('ignored', $currentValue).GetNetworkCredential().Password
+            }
+            elseif( $currentValue -is [TimeSpan] )
+            {
+                if( $expectedValue -match '^\d+$' )
+                {
+                    $expectedValue = [TimeSpan]::New([long]$expectedValue)
+                }
+                else
+                {
+                    $expectedValue = [TimeSpan]$expectedValue
+                }
+            }
+
+            $currentValue |
+                Should -Be $expectedValue -Because "should $($setMsg)$($asDefaultsMsg) $($attr.Name) to $($becauseMsg) value"
+        }
+    }
+}
+
+Describe 'Set-CIisWebsiteLimit' {
+    BeforeAll {
+        Start-W3ServiceTestFixture
+        Install-CIisAppPool -Name 'Set-CIisWebsiteLimit'
+    }
+
+    AfterAll {
+        Uninstall-CIisAppPool -Name 'Set-CIisWebsiteLimit'
+        Complete-W3ServiceTestFixture
+    }
+
+    BeforeEach {
+        $script:siteName = "Set-CIisWebsiteLimit$($script:testNum)"
+        $script:testNum++
+        Set-CIisWebsiteLimit -AsDefaults @script:defaultDefaults -Reset
+        Install-CIisWebsite -Name $script:siteName -PhysicalPath (New-TestDirectory) -AppPoolName 'Set-CIisWebsiteLimit'
+    }
+
+    AfterEach {
+        Uninstall-CIisWebsite -Name $script:siteName
+        Set-CIisWebsiteLimit -AsDefaults @script:defaultDefaults -Reset
+    }
+
+    It 'should set and reset all values' {
+        $infos = @()
+        Set-CIisWebsiteLimit -SiteName $script:siteName @script:nonDefaultArgs -Reset -InformationVariable 'infos'
+        $infos | Should -Not -BeNullOrEmpty
+        ThenHasValues $script:nonDefaultArgs
+
+        # Make sure no information messages get written because no changes are being made.
+        Set-CIisWebsiteLimit -SiteName $script:siteName @script:nonDefaultArgs -Reset -InformationVariable 'infos'
+        $infos | Should -BeNullOrEmpty
+        ThenHasValues $script:nonDefaultArgs
+
+        Set-CIisWebsiteLimit -SiteName $script:siteName -Reset
+        ThenHasDefaultValues
+    }
+
+    It 'should support WhatIf when updating all values' {
+        Set-CIisWebsiteLimit -SiteName $script:siteName @script:nonDefaultArgs -Reset -WhatIf
+        ThenHasDefaultValues
+    }
+
+    It 'should support WhatIf when resetting all values back to defaults' {
+        Set-CIisWebsiteLimit -SiteName $script:siteName -Reset @script:nonDefaultArgs
+        ThenHasValues $script:nonDefaultArgs
+        Set-CIisWebsiteLimit -SiteName $script:siteName -Reset -WhatIf
+        ThenHasValues $script:nonDefaultArgs
+    }
+
+    It 'should change values and not reset to defaults' {
+        Set-CIisWebsiteLimit -SiteName $script:siteName @script:nonDefaultArgs -ErrorAction Ignore
+        ThenHasValues $script:nonDefaultArgs
+
+        $someArgs = @{
+            'connectionTimeout' = '00:00:30';
+            'maxUrlSegments' = 8;
+        }
+        Set-CIisWebsiteLimit -SiteName $script:siteName @someArgs
+        ThenHasValues $someArgs -OrValues $script:nonDefaultArgs
+    }
+
+    It 'should change default settings' {
+        Set-CIisWebsiteLimit -AsDefaults @script:nonDefaultArgs -Reset
+        ThenDefaultsSetTo $script:nonDefaultArgs
+
+        $someArgs = @{
+            'maxBandwidth' = 268435455;
+            'maxConnections' = 134217727;
+        }
+        Set-CIisWebsiteLimit -AsDefaults @someArgs
+        ThenDefaultsSetTo $someArgs -OrValues $script:nonDefaultArgs
+    }
+}